moved styles into individual component so that they are self contained
