+++ samples/widgets/CpuUsage/recharts-style.css
