+++ samples/widgets/TotalUpdates/src/recharts-style.css
@@ -1,0 +1,3 @@
+.recharts-legend-item-text {
+    color: #5d6e77;
+}
